convert the last two dropdowns and uninstall @radix-ui/react-dropdown-menu
