+++ app/components/TopBarPicker.tsx
@@ -25,7 +25,7 @@ import {
 import { useCurrentUser } from '~/layouts/AuthenticatedLayout'
 import { PAGE_SIZE } from '~/table/QueryTable'
 import { buttonStyle } from '~/ui/lib/Button'
-import * as DropdownMenu from '~/ui/lib/DropdownMenu2'
+import * as DropdownMenu from '~/ui/lib/DropdownMenu'
 import { Identicon } from '~/ui/lib/Identicon'
 import { Wrap } from '~/ui/util/wrap'
 import { pb } from '~/util/path-builder'
